+++ src/Expression/Statement/When.php
@@ -8,14 +8,14 @@ use Yiisoft\Db\Query\QueryInterface;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 
 /**
- * Represents the condition and the result of a WHEN clause in a SQL CASE statement.
+ * Represents the condition and the result of a `WHEN` clause in a SQL `CASE` statement.
  *
- * @see CaseExpression
+ * @see CaseX
  */
-final class WhenClause
+final class When
 {
     /**
-     * @param mixed $condition The condition for the WHEN clause:
+     * @param mixed $condition The condition for the `WHEN` clause:
      * - `string` is treated as a SQL expression;
      * - `array` is treated as a condition to check, see {@see QueryInterface::where()};
      * - other values will be converted to their string representation using {@see QueryBuilderInterface::buildValue()}.