+++ tests/AbstractQueryBuilderTest.php
@@ -16,7 +16,7 @@ use Yiisoft\Db\Exception\Exception;
 use InvalidArgumentException;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
-use Yiisoft\Db\Expression\Statement\CaseExpression;
+use Yiisoft\Db\Expression\Statement\CaseX;
 use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\Expression\ExpressionBuilderInterface;
 use Yiisoft\Db\Expression\ExpressionInterface;
@@ -2418,9 +2418,9 @@ abstract class AbstractQueryBuilderTest extends TestCase
         Assert::arraysEquals($expectedParams, $params);
     }
 
-    #[DataProviderExternal(QueryBuilderProvider::class, 'caseExpressionBuilder')]
-    public function testCaseExpressionBuilder(
-        CaseExpression $case,
+    #[DataProviderExternal(QueryBuilderProvider::class, 'caseXBuilder')]
+    public function testCaseXBuilder(
+        CaseX $case,
         string $expectedSql,
         array $expectedParams,
         string|int $expectedResult,
@@ -2432,20 +2432,6 @@ abstract class AbstractQueryBuilderTest extends TestCase
 
         $this->assertSame($expectedSql, $qb->buildExpression($case, $params));
         $this->assertEquals($expectedParams, $params);
-    }
-
-    public function testCaseExpressionBuilderEmpty(): void
-    {
-        $db = $this->getConnection();
-        $qb = $db->getQueryBuilder();
-
-        $params = [];
-        $case = new CaseExpression();
-
-        $this->expectException(InvalidArgumentException::class);
-        $this->expectExceptionMessage('The CASE expression must have at least one WHEN clause.');
-
-        $qb->buildExpression($case, $params);
     }
 
     #[DataProviderExternal(QueryBuilderProvider::class, 'lengthBuilder')]